requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorclass==2.2.2
Django==4.2.1
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pip-upgrader==1.4.15
platformdirs==3.5.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
terminaltables==3.1.10
urllib3==2.0.2
